Сделаны основные таблицы и способы взаимодействия с ними

+++ src/lib/Tables/Table.ts
@@ -1,5 +1,5 @@
 import { PoolConnection } from "mariadb";
-import { ITable, ITableConfig, ITablePage, ITableJoin } from "../../types";
+import { ITable, ITableConfig, ITablePage, ITableJoin, SQL } from "../../types";
 import {
 	parseCreateTable,
 	toValidSQLKeys,
@@ -20,13 +20,16 @@ export class Table implements ITable {
 
 	public async init(connection: PoolConnection | null) {
 		this.connection = connection;
+
 		const intiSQL: string = parseCreateTable(this.config);
+
 		await this.connection!.query(intiSQL);
 	}
 
 	public async insertData<Request extends Object>(params: Request) {
 		const fields: string = toValidSQLKeys(params);
 		const values: string = toValidSQLValues(params);
+
 		await this.connection?.query(
 			`INSERT ${this.config.table}(${fields}) VALUES(${values});`
 		);
@@ -37,8 +40,8 @@ export class Table implements ITable {
 		filters?: Object,
 		join?: ITableJoin[]
 	) {
-		let where: string = "";
-		let association: string = "";
+		let where: SQL = "";
+		let association: SQL = "";
 
 		if (typeof filters !== "undefined") {
 			where = toValidSQLWhere(filters);
@@ -48,8 +51,8 @@ export class Table implements ITable {
 			association = parseJoinTables(join);
 		}
 
-		const start = (page.page - 1) * page.countOnPage;
-		const end = page.page * page.countOnPage;
+		const start: number = (page.page - 1) * page.countOnPage;
+		const end: number = page.page * page.countOnPage;
 
 		const response: Response[] = await this.connection?.query(
 			`SELECT * FROM ${this.config.table} ${where} ${association} LIMIT ${start}, ${end};`
@@ -59,20 +62,28 @@ export class Table implements ITable {
 	}
 
 	public async deleteData<Filters extends Object>(filters: Filters) {
-		const where = toValidSQLWhere(filters);
+		if (Object.getOwnPropertyNames(filters).length === 0) {
+			throw new Error("filters must have any property");
+		}
+		const where: SQL = toValidSQLWhere(filters);
+
 		await this.connection?.query(`DELETE FROM ${this.config.table} ${where};`);
 	}
 
 	public async updateData<Values extends Object, Filters extends Object>(
 		newValues: Values,
-		filters?: Filters
+		filters: Filters
 	) {
-		let where: string = "";
-		if (typeof filters !== "undefined") {
-			where = toValidSQLWhere(filters);
+		if (
+			Object.getOwnPropertyNames(newValues).length === 0 ||
+			Object.getOwnPropertyNames(filters).length === 0
+		) {
+			throw new Error("newValues and Filter must have any property");
 		}
+		console.log(Object.getOwnPropertyNames(newValues));
+		const update: SQL = parseSetParams(newValues);
+		const where: SQL = toValidSQLWhere(filters);
 
-		const update: string = parseSetParams(newValues);
 		await this.connection?.query(
 			`UPDATE ${this.config.table} SET ${update} ${where};`
 		);
